Fix election positions

+++ humhub/plugins/election/controllers/PositionController.php
@@ -8,6 +8,7 @@ use humhub\modules\election\permissions\CreateElection;
 use humhub\modules\space\models\Space;
 use Yii;
 use yii\web\ForbiddenHttpException;
+use yii\web\HttpException;
 use yii\web\NotFoundHttpException;
 
 class PositionController extends ContentContainerController
@@ -37,6 +38,8 @@ class PositionController extends ContentContainerController
         $this->requireAdmin();
         $model = new ElectionPosition();
         $model->space_id = $this->contentContainer->id;
+        $model->is_default = 0;
+        $model->is_active = 1;
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             $this->view->saved();
@@ -71,7 +74,24 @@ class PositionController extends ContentContainerController
         $this->forcePostRequest();
 
         $model = $this->findPosition($id);
+
+        if ($model->isDefault()) {
+            throw new HttpException(403, Yii::t('ElectionModule.base', 'Default positions cannot be deleted. You can disable them instead.'));
+        }
+
         $model->delete();
+        $this->view->saved();
+        return $this->redirect($this->contentContainer->createUrl('/election/position/index'));
+    }
+
+    public function actionToggle($id)
+    {
+        $this->requireAdmin();
+        $this->forcePostRequest();
+
+        $model = $this->findPosition($id);
+        $model->is_active = $model->isActive() ? 0 : 1;
+        $model->save(false);
 
         $this->view->saved();
         return $this->redirect($this->contentContainer->createUrl('/election/position/index'));
